feat: :sparkles: Completed implementation of Teacher's progress review and fixed some bugs

+++ database/migrations/2024_09_14_173944_create_user_activities_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->primary(['activityId', 'userId']);
             $table->foreign('activityId')->references('activityId')->on('activities');
             $table->foreign('userId')->references('userId')->on('users');
-            $table->integer('score')->nullable();
+            $table->decimal('score', 8, 2)->nullable();
             $table->String('answer')->nullable();
             $table->timestamps();
         });
